Revert "refactor: PasswordEncoder 이동"
This reverts commit d18ccf0f57985d0bc1997629a7f99a93ad7e2f0c.
