+++ src/main/java/shopping/infrastructure/PasswordEncoder.java
@@ -1,4 +1,4 @@
-package shopping.domain.user;
+package shopping.infrastructure;
 
 public interface PasswordEncoder {
     String encode(final String password);
